requirements.txt
Django==1.6.2
Rdio==0.2
South==0.8.4
argparse==1.2.1
distribute==0.6.34
django-filter==0.7
django-localflavor==1.0
django-model-utils==2.0.3
django-phonenumber-field==0.2a3
django-twilio==0.5
djangorestframework==2.3.13
httplib2==0.8
six==1.6.1
twilio==3.6.6
wsgiref==0.1.2
xmltodict==0.8.7
